Widen IE64 JIT block infrastructure to uint64 PCs
Phase 3 of the IE64 full-64-bit JIT plan: widen every block-cache,
chain-slot, RTS-cache, region, and emitter PC type from uint32 to
uint64 so that future phases (block fetch, helper callouts, full 64-bit
execution) have a coherent type surface to build on.

Type changes (uint32 → uint64):
  - JITBlock.startPC, endPC, coveredRanges
  - chainSlot.targetPC, ie64PendingChainSlot.targetPC
  - JITContext.RTSCache0..3PC (RTS-cache slot PC entries)
  - ie64Region.entryPC, blockPCs

API signature changes:
  - CodeCache.Get/Put, PatchChainsTo, PatchChainsToScoped
  - CodeCache.InvalidateRange, UnpatchChainsInRange
  - compileBlock, compileBlockMMU, scanBlock, scanBlockWithLimit
  - detectBackwardBranches, ie64ResolveTerminatorTarget
  - ie64FormRegion, analyzeJSRLeafFusion
  - All emitter instrPC / blockStartPC / blockPC params on AMD64 + ARM64

Exact MMU cache key:
  CodeCache previously hashed (ptbr, virtPC) through a lossy golden-ratio
  multiplicative scheme that could collide distinct {ptbr, pc} pairs.
  Replace with a dedicated mmuBlocks map keyed on an ie64CacheKey
  struct so high-PTBR or high-vPC tasks cannot alias each other's
  cached blocks. Add GetMMU / PutMMU; extend InvalidateRange,
  PatchChainsTo, PatchChainsToScoped, and UnpatchChainsInRange to walk
  both the non-MMU and MMU maps.

AMD64 RTS cache:
  Widen the 4-entry MRU probe to 64-bit CMP (new amd64ALU_reg_mem_cmp
  helper) against the now-uint64 RTSCache*PC fields. Widen the empty-
  slot zero guard to TEST RAX, RAX. Remove the Phase 2 high-bit bypass
  — the bypass became redundant once the cache PC width matches the
  popped return address width.

ARM64 JSR target:
  emitJSR previously double-cast through uint32 (uint64(uint32(...)))
  on the PC-relative sum, aliasing high-address calls to their low 32
  bits. Drop the intermediate uint32 cast so the chain exit / RetPC
  carry the full 64-bit target, matching the AMD64 JSR and the ARM64
  BRA/Bcc/JMP paths fixed in earlier phases.

Non-IE64 backends (6502 / M68K / Z80 / x86) keep their narrower
internal PC types (uint16 / uint32 / uint32) and cast at the
CodeCache/JITBlock boundary. Their RTSCache*PC ctx fields (separate
from JITContext) remain uint32; the boundary casts go uint32(block.
startPC) where the JITBlock field is uint64.

Tests:
  - jit_cache_widen_test.go covers high-PC block lookup, InvalidateRange
    over a high PC, UnpatchChainsInRange recognising a high doomed
    target, MMU exact-key isolation across two PTBRs (incl. >4 GiB
    PTBRs that the legacy hash could collide), MMU-scoped
    InvalidateRange, and ie64ResolveTerminatorTarget sign-extension at
    a high instrPC with a negative imm32.
  - jit_ret_channel_test.go adds TestJIT_JSR_PCRelTarget_FullWidth
    pinning the JSR PC-relative target width. The compileAndRun rig
    fixes instrPC at PROG_START so the sum cannot overflow uint32 yet;
    the >4 GiB overflow case requires Phase 4's high-PC block fetch.

Verification:
  - go test -tags headless -count=1 ./... — only the pre-existing
    TestJIT6502_DirectPageBitmap_Ranges (6502 backend, unrelated) fails
  - ARM64 via qemu-aarch64 (binfmt-registered) — only the pre-existing
    TestJIT_FCVTFI_Saturate fails
  - go test -race — clean across IE64 / AMD64 / cache test surface

+++ jit_6502_emit_amd64.go
@@ -2745,14 +2745,14 @@ done:
 	var slots []chainSlot
 	for _, ce := range chainExits {
 		slots = append(slots, chainSlot{
-			targetPC:  ce.targetPC,
+			targetPC:  uint64(ce.targetPC),
 			patchAddr: addr + uintptr(ce.jmpDispOffset),
 		})
 	}
 
 	return &JITBlock{
-		startPC:    uint32(startPC),
-		endPC:      uint32(lastByte) + 1, // first byte after block (for cache invalidation)
+		startPC:    uint64(startPC),
+		endPC:      uint64(lastByte) + 1, // first byte after block (for cache invalidation)
 		instrCount: len(instrs),
 		execAddr:   addr,
 		execSize:   len(code),
