+++ jit_6502_exec.go
@@ -183,7 +183,7 @@ func (cpu *CPU_6502) ExecuteJIT6502() {
 		}
 
 		// Try cached block
-		block := cpu.jitCache.Get(uint32(pc))
+		block := cpu.jitCache.Get(uint64(pc))
 		if block == nil {
 			// Scan and potentially compile a new block
 			instrs := p65ScanTurboBlock(cpu, mem, pc, memSize)
@@ -308,7 +308,7 @@ func (cpu *CPU_6502) ExecuteJIT6502() {
 		if block.chainEntry != 0 {
 			ctx.RTSCache1PC = ctx.RTSCache0PC
 			ctx.RTSCache1Addr = ctx.RTSCache0Addr
-			ctx.RTSCache0PC = block.startPC
+			ctx.RTSCache0PC = uint32(block.startPC)
 			ctx.RTSCache0Addr = block.chainEntry
 		}
 
@@ -344,8 +344,8 @@ func (cpu *CPU_6502) ExecuteJIT6502() {
 			lo := page << 8
 			hi := lo + 256
 			// Unpatch chain slots targeting invalidated range, then remove blocks
-			cpu.jitCache.UnpatchChainsInRange(lo, hi)
-			cpu.jitCache.InvalidateRange(lo, hi)
+			cpu.jitCache.UnpatchChainsInRange(uint64(lo), uint64(hi))
+			cpu.jitCache.InvalidateRange(uint64(lo), uint64(hi))
 			// Conservative: leave codePageBitmap stale (false positives are safe).
 			// Stale entries cleared on full ExecMem exhaustion reset.
 			ctx.NeedInval = 0
